[develop] Split Util class in another module
